initial build and deploy
tested on several garments. seems to work well for organizing the production info into a hermetic layer and remove any existing cut line callout text. then adds new callouts with names from the layers panel and a flag for the order number to use a replace()

then after the whole prepress layer is processed, the rest of the styles in the source folder get opened and the existing prod info stuff is removed, then the source file prod info groups are duplicated into the current document which is then saved into a batched files folder.